rename EditableElements, create Li element
